requirements.txt
black==25.1.0
click==8.2.1
colorama==0.4.6
isort==6.0.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
PyAudio==0.2.14
Pygments==2.19.1
pyzmq==26.4.0
rich==14.0.0
textual==3.2.0
typing_extensions==4.13.2
uc-micro-py==1.0.3
